refactor(textareafield): use inputlabel and re-export as subcomp (#1701)
* refactor(textareafield): use inputlabel and re-export as subcomp

* feat(textareafield)!: remove top level subcomponent
