--- src/components/TextArea/TextArea.module.css
@@ -1,16 +1,0 @@
-@import '../../design-tokens/mixins.css';
-
-/*------------------------------------*\
-    # TEXTAREA 
-\*------------------------------------*/
-
-/**
- * Default input styles 
- */
-.textarea {
-  @mixin inputStyles;
-}
-
-.textarea--disabled:focus {
-  outline: none;
-}
